Add files via upload
Script to implement multiple choice. More features can be added to add modifications to a score script, health damage script or another check. I have also included the json to see how things are set up. This is just a basis for a start. This can also be modified for dialogue selection as a multiple choice option.